main modified

+++ main.py
@@ -7,16 +7,21 @@ from dash import Dash
 from src import graph
 
 if __name__ == "__main__":
-    try :
-        raw_data=pd.read_json("data/raw_data.json")
-        print('-----Finished to load raw_data-----')
-    except:
+    force_writing=False
+    if not force_writing:  
+        try :
+            raw_data=pd.read_json("data/raw_data.json")
+            print('-----Finished to load raw_data-----')
+        except Exception as e:
+            print(e)
+            raw_data=None
+    if raw_data.iloc[0,0]==None or force_writing :
         db_conn = sqlite3.connect("data/scrap.db")
         db_cursor = db_conn.cursor()
         raw_data = data.db_to_dataframe(db_cursor)
         raw_data.to_json('data/raw_data.json')
         print('-----Raw_data registered-----')
-    graph_dict=data_analysis.graph_dict_generate(raw_data.head(1000),True)
+    graph_dict=data_analysis.graph_dict_generate(raw_data.head(1000),force_writing)
     app = Dash(__name__)
     app.layout =graph.dash_graph(graph_dict)
     app.run(debug=True)
